requirements.txt
numpy
matplotlib
Pillow
imageio
